Clean up now unused VariantValue code
