+++ src/common/types/variant/CMakeLists.txt
@@ -1,5 +1,5 @@
 add_library_unity(duckdb_common_variant OBJECT variant.cpp variant_iterator.cpp
-                  variant_value.cpp variant_value_convert.cpp)
+                  variant_value_convert.cpp)
 set(ALL_OBJECT_FILES
     ${ALL_OBJECT_FILES} $<TARGET_OBJECTS:duckdb_common_variant>
     PARENT_SCOPE)
