+++ src/include/duckdb/common/enum_util.hpp
@@ -542,8 +542,6 @@ enum class VariantLogicalType : uint8_t;
 
 enum class VariantStatsShreddingState : uint8_t;
 
-enum class VariantValueType : uint8_t;
-
 enum class VectorBufferType : uint8_t;
 
 enum class VectorType : uint8_t;
@@ -1327,9 +1325,6 @@ const char* EnumUtil::ToChars<VariantLogicalType>(VariantLogicalType value);
 
 template<>
 const char* EnumUtil::ToChars<VariantStatsShreddingState>(VariantStatsShreddingState value);
-
-template<>
-const char* EnumUtil::ToChars<VariantValueType>(VariantValueType value);
 
 template<>
 const char* EnumUtil::ToChars<VectorBufferType>(VectorBufferType value);
@@ -2123,9 +2118,6 @@ VariantLogicalType EnumUtil::FromString<VariantLogicalType>(const char *value);
 
 template<>
 VariantStatsShreddingState EnumUtil::FromString<VariantStatsShreddingState>(const char *value);
-
-template<>
-VariantValueType EnumUtil::FromString<VariantValueType>(const char *value);
 
 template<>
 VectorBufferType EnumUtil::FromString<VectorBufferType>(const char *value);
